Task 3: Multi-modal Chatbot/requirements_multimodal_chatbot.txt
streamlit==1.35.0
Pillow==10.3.0
opencv-python==4.9.0.80
requests==2.31.0
google-generativeai==0.4.1
numpy==1.26.4
matplotlib==3.8.4